% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualizar_utilidades3F.R
\name{actualizar_utilidades3F}
\alias{actualizar_utilidades3F}
\title{actualizar_utilidades3F}
\usage{
actualizar_utilidades3F(ref = "estable", ...)
}
\arguments{
\item{ref}{Rama o tag a instalar. Por defecto usa "estable".}

\item{...}{Argumentos adicionales que se pasan a `devtools::install_github()`.}
}
\value{
Mensaje de instalación o error.
}
\description{
Reinstala el paquete `utilidades3F` desde el repositorio oficial de GitHub.
}
\examples{
actualizar_utilidades3F()  # Instala desde rama "estable"
actualizar_utilidades3F(ref = "otra-rama")  # Instala desde una rama alternativa
}
