% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBA.R
\name{ARBA_inventario}
\alias{ARBA_inventario}
\title{ARBA_inventario}
\usage{
ARBA_inventario()
}
\value{
devuelve un listado de las capa disponibles para descargar de los geoservicios de ARBA
}
\description{
ARBA_inventario
}
