% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estructurar_directorio.R
\name{estructurar_directorio}
\alias{estructurar_directorio}
\title{estructurar_directorio}
\usage{
estructurar_directorio(
  verificar_proyecto = TRUE,
  nombre_insumos = "insumos",
  nombre_productos = "productos"
)
}
\arguments{
\item{verificar_proyecto}{TRUE o FALSE para activar la verificación de proyecto.}

\item{nombre_insumos}{string con el nombre de la carpeta de insumos.}

\item{nombre_productos}{string con el nombre de la carpeta de productos.}
}
\value{
Verificará si existe un proyecto en el directorio de trabajo actual y
 creará las carpetas de insumos y productos si no existen.
}
\description{
Permite validar si existe un proyecto en el directorio de trabajo actual
y verificar que exista una subcarpeta de insumos y otra de productos en el mismo.
En caso de que las carpetas no existan las crea.
}
\examples{
estructurar_directorio(verificar_proyecto = TRUE,
 nombre_insumos = "insumos",
 nombre_productos = "productos")
}
