% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBA.R
\name{inventario_ARBA}
\alias{inventario_ARBA}
\title{inventario_ARBA}
\usage{
inventario_ARBA()
}
\value{
devuelve un listado de las capa disponibles para descargar de los geoservicios de ARBA
}
\description{
inventario_ARBA
}
