% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar_calles.R
\name{normalizar_calles}
\alias{normalizar_calles}
\title{normalizar_calles}
\usage{
normalizar_calles(df, nombre_calles)
}
\arguments{
\item{df}{base de datos}

\item{nombre_calles}{columna que contenga los nombres de calles}
}
\value{
devuelve la base de datos con una columna adicional con los nombres de calles normalizados
}
\description{
Permite normalizar los nombres de calles de una base de datos para el municipio de Tres de Febrero
}
\examples{
nombre_normalizados <- normalizar_calles(df, nombre_calles = "calle_nombre");
}
