% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar_localidades.R
\name{normalizar_localidades}
\alias{normalizar_localidades}
\title{Permite normalizar y corregir los nombres de las localidades en base a la superposición geométrica}
\usage{
normalizar_localidades(df = NULL, localidades = NULL)
}
\arguments{
\item{df}{Un objeto `sf` (capa) a normalizar.}

\item{localidades}{Un objeto `sf` con las localidades de referencia. Si es `NULL` (por defecto), 
se descargará automáticamente la capa "localidades" del servidor.}
}
\value{
Devuelve un objeto `sf` con la geometría unida espacialmente.
}
\description{
Permite normalizar y corregir los nombres de las localidades en base a la superposición geométrica
}
\examples{
\dontrun{
capa <- obtener_capa(nombre_de_capa = "puntos_interes")
capa_normalizada <- normalizar_localidades(df = capa)
}
}
