% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBA.R
\name{obtener_capa_ARBA}
\alias{obtener_capa_ARBA}
\title{obtener_capa_ARBA}
\usage{
obtener_capa_ARBA(nombre_de_capa)
}
\arguments{
\item{nombre_de_capa}{nombre de la capa a descargar}
}
\value{
devuelve la capa deseada, en caso de no conocer el nombre utilizar la función de inventario_obtener_capa_ARBA()
}
\description{
obtener_capa_ARBA
}
\examples{
capa <- obtener_capa_ARBA(nombre_de_capa = "Parcela");
}
