% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validar_paquetes.R
\name{validador_paquetes}
\alias{validador_paquetes}
\title{validador_paquetes}
\usage{
validador_paquetes(paquetes)
}
\arguments{
\item{paquetes}{vector con los paquetes a instalar}
}
\description{
permite validar si los paquetes necesarios en un script se encuentran instalados
en caso de no estar instalados, se instalan
}
\examples{
validador_paquetes(paquetes);
}
