% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver_tabla.R
\name{ver_tabla}
\alias{ver_tabla}
\title{Visualizar datos sin notación científica}
\usage{
ver_tabla(x, digits = NULL, ...)
}
\arguments{
\item{x}{Un data.frame o matriz a visualizar.}

\item{digits}{Entero (opcional). Cantidad de decimales fijos a mostrar.}

\item{...}{Argumentos adicionales pasados a \code{utils::View} (ej. \code{title}).}
}
\value{
Abre el visor de datos.
}
\description{
Abre el visor de datos de R (View) formateando previamente las columnas numéricas
para evitar la notación científica. Es útil para visualizar identificadores largos.
}
