% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autenticar_geoportal.R
\name{autenticar_geoportal}
\alias{autenticar_geoportal}
\title{autenticar_geoportal}
\usage{
autenticar_geoportal(
  client_id = NULL,
  client_secret = NULL,
  guardar = TRUE,
  usar_cache = TRUE,
  usar_keyring = FALSE
)
}
\arguments{
\item{client_id}{ID del cliente OAuth2. Se intenta primero desde `.Renviron`, luego desde `keyring`, y por último del argumento.}

\item{client_secret}{Clave secreta OAuth2.}

\item{guardar}{Si TRUE, guarda el token generado y las credenciales en `.Renviron` si se pasaron como argumento.}

\item{usar_cache}{Si TRUE, reutiliza el token guardado si tiene menos de 12 horas.}

\item{usar_keyring}{Si TRUE, intenta recuperar las credenciales desde el sistema seguro de `keyring`.}
}
\value{
Un objeto de clase `Token2.0` con credenciales OAuth2.
}
\description{
Inicia sesión en el GeoPortal de Tres de Febrero mediante OAuth2 y devuelve un token de acceso válido.
}
