% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autenticar_geoportal.R
\name{autenticar_geoportal}
\alias{autenticar_geoportal}
\title{autenticar_geoportal}
\usage{
autenticar_geoportal(
  client_id = NULL,
  client_secret = NULL,
  guardar = TRUE,
  usar_cache = TRUE,
  ignorar_SSL = FALSE
)
}
\arguments{
\item{client_id}{ID del cliente OAuth2. Si no está en `.Renviron`, se puede pasar como argumento.}

\item{client_secret}{Clave secreta OAuth2. Si no está en `.Renviron`, se puede pasar como argumento.}

\item{guardar}{Si TRUE, guarda el token y también las credenciales en `.Renviron` si no están.}

\item{usar_cache}{Si TRUE, intenta reutilizar el token guardado si tiene menos de 12 horas.}

\item{ignorar_SSL}{Si TRUE, ignora los errores SSL (por ejemplo, certificados vencidos o auto-firmados).}
}
\value{
Un objeto de clase `Token2.0` con credenciales OAuth2.
}
\description{
Inicia sesión en el GeoPortal de Tres de Febrero mediante OAuth2 y devuelve un token de acceso válido.
}
