% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_capas.R
\name{autenticar_geoportal}
\alias{autenticar_geoportal}
\title{autenticar_geoportal}
\usage{
autenticar_geoportal(client_id = NULL, client_secret = NULL, guardar = TRUE)
}
\arguments{
\item{client_id}{ID del cliente OAuth2 registrado en el geoportal. No es necesario si el token ya fue guardado previamente.}

\item{client_secret}{Clave secreta del cliente OAuth2. No es necesario si el token ya fue guardado previamente.}

\item{guardar}{Si TRUE, guarda el token en un directorio persistente del usuario y lo reutiliza en sesiones siguientes.}
}
\value{
Un objeto de clase `Token2.0` con credenciales OAuth2.
}
\description{
Inicia sesión en el GeoPortal de Tres de Febrero mediante OAuth2 y devuelve un token de acceso válido.
}
\examples{
token <- autenticar_geoportal("mi_client_id", "mi_client_secret", guardar = TRUE)
token <- autenticar_geoportal()  # Reutiliza el token ya guardado
}
