% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurar_credenciales_db.R
\name{configurar_credenciales_db}
\alias{configurar_credenciales_db}
\title{Configura las credenciales para un servicio de base de datos en el keyring}
\usage{
configurar_credenciales_db(nombre_servicio)
}
\arguments{
\item{nombre_servicio}{El nombre del servicio (ej. "SSO_vista")}
}
\description{
Esta función solicita interactivamente al usuario que guarde el servidor,
el usuario y la contraseña de la BD para un servicio específico.
}
