% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_conexiones.R
\name{crear_conexiones}
\alias{crear_conexiones}
\title{Crea una o más conexiones a bases de datos}
\usage{
crear_conexiones(configs, file = "config.yml")
}
\arguments{
\item{configs}{Un vector de caracteres con los nombres de las
configuraciones a cargar.}

\item{file}{La ruta al archivo de configuración. Por defecto "config.yml".}
}
\value{
Una lista nombrada, donde cada elemento es un objeto `DBIConnection`.
}
\description{
Esta función toma uno o más nombres de configuración y devuelve
una lista nombrada de conexiones DBI.
}
