% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_capas.R
\name{inventario_capas}
\alias{inventario_capas}
\title{inventario_capas}
\usage{
inventario_capas(usar_autenticacion = FALSE)
}
\arguments{
\item{usar_autenticacion}{Lógico. Si TRUE, intenta cargar el token guardado del usuario para mostrar capas privadas.}
}
\value{
Un vector con los nombres de las capas disponibles.
}
\description{
Lista las capas disponibles en el geoportal. Ejecuta en entorno limpio si `usar_autenticacion = FALSE`.
}
