% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_capas.R
\name{inventario_capas}
\alias{inventario_capas}
\title{inventario_capas}
\usage{
inventario_capas(
  wfs = "https://geoportal.tresdefebrero.gob.ar/geoserver/ows",
  service_version = "1.1.0",
  pretty_output = TRUE
)
}
\value{
devuelve las capas disponibles
}
\description{
Nos devuelve un listado de las capas disponibles en el geoportal municipal
}
