% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar_localidades.R
\name{normalizar_localidades}
\alias{normalizar_localidades}
\title{obtener_capa}
\usage{
normalizar_localidades(df = NULL, localidades = localidades_oc)
}
\arguments{
\item{df}{nombre de la capa a descargar}

\item{localidades}{description}
}
\value{
devuelve la capa deseada, en caso de no conocer el nombre utilizar la función de inventario_capas
}
\description{
Permite normalizar y corregir los nombres de las localidades en base a la superposición geométrica
}
\examples{
capa <- obtener_capa(nombre_de_capa = "localidades");
}
