% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_capas.R
\name{obtener_capa}
\alias{obtener_capa}
\title{obtener_capa}
\usage{
obtener_capa(nombre_de_capa)
}
\arguments{
\item{nombre_de_capa}{nombre de la capa a descargar}
}
\value{
devuelve la capa deseada, en caso de no conocer el nombre utilizar la función de inventario_capas()
}
\description{
Permite descargar una capa del geoportal municipal ingresando su nombre
}
\examples{
capa <- obtener_capa(nombre_de_capa = "localidades");
}
