% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_inventario.R
\name{obtener_inventario}
\alias{obtener_inventario}
\title{obtener_inventario}
\usage{
obtener_inventario(
  usar_autenticacion = FALSE,
  limpiar_prefijo = TRUE,
  ignorar_SSL = FALSE,
  .interno = FALSE
)
}
\arguments{
\item{usar_autenticacion}{Lógico. Si TRUE, intenta cargar el token guardado del usuario para mostrar capas privadas.}

\item{limpiar_prefijo}{Lógico. Si TRUE, remueve el prefijo "geonode:" de los nombres de capa.}

\item{ignorar_SSL}{Lógico. Si TRUE, desactiva la verificación SSL (solo usar en entornos de desarrollo o servidores internos).}

\item{.interno}{No tocar. Usado internamente para evitar recursión infinita.}
}
\value{
Un vector con los nombres de las capas disponibles
}
\description{
Lista las capas disponibles en el geoportal.
}
