% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validar_paquetes.R
\name{validar_paquetes}
\alias{validar_paquetes}
\title{validar_paquetes}
\usage{
validar_paquetes(paquetes)
}
\arguments{
\item{paquetes}{vector con los paquetes a instalar}
}
\description{
permite validar si los paquetes necesarios en un script se encuentran instalados
en caso de no estar instalados, se instalan
}
\examples{
paquetes <- c("sf")
validar_paquetes(paquetes);
}
