% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validar_paquetes.R
\name{validar_paquetes}
\alias{validar_paquetes}
\title{Validar e instalar paquetes necesarios}
\usage{
validar_paquetes(paquetes)
}
\arguments{
\item{paquetes}{Vector de caracteres con los nombres de los paquetes.}
}
\description{
Verifica si una lista de paquetes está instalada. Si alguno falta, lo instala automáticamente.
Finalmente, carga los paquetes en la sesión actual.
}
\examples{
\dontrun{
  validar_paquetes(c("dplyr", "sf"))
}
}
